pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moodyduck"
version = "0.1.0"
authors = [{ name = "Kumi Mitterer", email = "kumify@kumi.email" }]
description = "A Django-based application for managing your life."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "django[argon2]",
    "django-colorfield",
    "django-multiupload",
    "django-storages",
    "django-polymorphic@git+https://github.com/jazzband/django-polymorphic.git@v4.11.1",
    "django-multiselectfield@git+https://git.private.coffee/kumi/django-multiselectfield.git",
    "django-multiupload@git+https://github.com/Chive/django-multiupload.git",
    "dbsettings",
    "django-filtersignals",
    "mysqlclient",
    "boto3",
    "argon2_cffi",
    "python-telegram-bot",
    "python-dateutil",
    "matrix-nio",
    "holoviews",
    "bokeh",
    "panel",
    "scipy",
    "channels",
    "django-autosecretkey",
    "mozilla-django-oidc",
    "python-gnupg",
    "django-crispy-forms",
    "crispy-bootstrap5",
    "django-annoying",
    "djangorestframework",
]

[project.scripts]
moodyduck = "moodyduck.manage:main"

[project.entry-points."moodyduck.gateways"]
telegram = "moodyduck.msgio.gateways.telegram:TelegramGateway"
matrix = "moodyduck.msgio.gateways.matrix:MatrixGateway"
ntfy = "moodyduck.msgio.gateways.ntfy:NtfyGateway"

[tool.hatch.metadata]
allow-direct-references = true

[project.urls]
"Homepage" = "https://git.private.coffee/kumi/kumify"
"Bug Tracker" = "https://git.private.coffee/kumi/kumify/issues"
"Source Code" = "https://git.private.coffee/kumi/kumify"
